refactor: Extract HomePage to a separate file"
