+++ lib/main.dart
@@ -4,12 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:audio_session/audio_session.dart';
 import 'package:logging/logging.dart';
 import 'package:zakstreamer/main_screen.dart';
-import 'package:zakstreamer/statistics_page.dart';
 import 'package:zakstreamer/statistics_service.dart';
-import 'package:zakstreamer/widgets/play_button.dart';
-import 'package:zakstreamer/widgets/now_playing_widget.dart';
 import 'page_manager.dart';
-import 'schedule_page.dart';
 import 'service_locator.dart';
 import 'package:flutter/services.dart';
 import 'notifications.dart';
@@ -102,58 +98,6 @@ class _ZakStreamerState extends State<ZakStreamer> {
       title: 'Żak Streamer',
       theme: ThemeData.dark(),
       home: const MainScreen(),
-    );
-  }
-}
-
-class HomePage extends StatelessWidget {
-  const HomePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Padding(
-          padding: EdgeInsets.symmetric(vertical: 36, horizontal: 12),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              const SizedBox(height: 140, child: NowPlayingWidget()),
-              const Text('Wciśnij Kropkę, aby włączyć alternatywę.'),
-              const PlayButton(),
-              TextButton(
-                child: const Text(
-                  'POKAŻ RAMÓWKĘ',
-                  style: TextStyle(color: Colors.tealAccent),
-                ),
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => const SchedulePage(),
-                    ),
-                  );
-                },
-              ),
-              TextButton(
-                child: const Text(
-                  'STATYSTYKI',
-                  style: TextStyle(color: Colors.white),
-                ),
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => const StatisticsPage(),
-                    ),
-                  );
-                },
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
